:art: fix: Change state management from tracings dropdown on sidebar
This change ensures that the object will be redrawn with newer values. The problem was that in the moment that dropdown is changed, the constructor call is not called again, preserving the object. When setState is called, the constructor is called with the same arguments, in this way, Flutter will not recreate the object.

So, to fix this problem, the constructor for the object must be updated by it's caller